refactor: consolidate smaller go files
